requirements.txt
matplotlib>=3.5.0
python-telegram-bot>=20.0

